% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Celligner.R
\name{Celligner}
\alias{Celligner}
\title{Celligner: tumor and cell line model normalized RNASeq data alignment}
\usage{
Celligner(tumor, cell)
}
\arguments{
\item{tumor}{a matrix or data frame of normalized tumor gene expression on log scale (log2-TPM preferred) with rows for genes and columns for samples.}

\item{cell}{a matrix or data frame of normalized cell line gene expression on log scale (log2-TPM preferred) with rows for genes and columns for samples.}
}
\value{
a Seurat object with aligned data included. \code{Seurat::GetAssayData()} can be used to extract the aligned dataset.
}
\description{
A two-step machine learning method to make normalized RNASeq data (log2 TPM preferred) from cell lines and tumors comparable.
To gain the benefits from the large sample size, it is recommended to merge the interested tumor samples with the public datasets such as TCGA.
The parameters are set as default and consistent with the Celligner paper, and the users can refer to the origional paper for more details.
}
\references{
Warren, Allison, Yejia Chen, Andrew Jones, Tsukasa Shibue, William C. Hahn, Jesse S. Boehm, Francisca Vazquez, Aviad Tsherniak, and James M. McFarland. 2021. “Global Computational Alignment of Tumor and Cell Line Transcriptional Profiles.” Nature Communications 12 (1): 22. https://doi.org/10.1038/s41467-020-20294-x.
}
