% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_prepare.R
\name{create_CASCAM}
\alias{create_CASCAM}
\title{Create a CASCAM object}
\usage{
create_CASCAM(tumor_aligned_data, tumor_label, camod_aligned_data, info_object)
}
\arguments{
\item{tumor_aligned_data}{A data frame of aligned tumor gene expression with rows for genes and columns for samples.}

\item{tumor_label}{A vector of strings indicating the labels of the tumor samples.}

\item{camod_aligned_data}{A data.frame of aligned cancer model gene expression with rows for genes and columns for samples.}

\item{info_object}{An \code{InformativeGenes} object for the study.}
}
\value{
A \code{CASCAM} object.
}
\description{
This function creates a \code{CASCAM} object for the downstream analysis. The data sets of tumor aligned gene expression data with the tumor subtype labels and
cancer model aligned gene expression data are needed as the input. Celligner is highly recommended to align the tumor and cancer model data.
}
