% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_genome_analysis.R
\name{genome_selection}
\alias{genome_selection}
\title{Genome-wide pre-selection analysis}
\usage{
genome_selection(object, R = 1000)
}
\arguments{
\item{object}{A \code{CASCAM} object.}

\item{R}{The number of bootstrap replicates. R = 1000 by default.}
}
\value{
A \code{CASCAM} object with genome-wide pre-selection related slots.
}
\description{
Genome-wide pre-selection analysis
}
\examples{
\dontrun{
gene_info <- create_InformativeGenes(tumor_ct, tumor_label2, "ILC")
CASCAM_eg <- create_CASCAM(tumor_aligned, tumor_label, camod_aligned, gene_info)
CASCAM_eg <- sda_model(CASCAM_eg)
CASCAM_eg <- genome_selection(CASCAM_eg)
}
}
