% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_prepare.R
\docType{class}
\name{InformativeGenes-class}
\alias{InformativeGenes-class}
\alias{InformativeGenes}
\title{The InformativeGenes class}
\value{

}
\description{
The InformativeGenes class
}
\section{Slots}{

\describe{
\item{\code{DEA}}{A data frame of the results of differential gene expression analysis.}

\item{\code{GSEA}}{A data frame of the results of gene set enrichment analysis.}

\item{\code{interested_subtype}}{A string indicating the interested tumor subtype.}
}}

