% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_pathway_analysis.R
\name{pathview_analysis}
\alias{pathview_analysis}
\title{Pathview of the interested camod in interested pathway}
\usage{
pathview_analysis(object, interested_camod, pathway_name)
}
\arguments{
\item{object}{A \code{CASCAM} object.}

\item{interested_camod}{A string of one interested cancer model.}

\item{pathway_name}{The interested pathway, which should be within the KEGG and Hallmark pathways.}
}
\description{
Pathview of the interested camod in interested pathway
}
