% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_pathway_analysis.R
\name{pathway_analysis}
\alias{pathway_analysis}
\title{Pathway specific selection analysis}
\usage{
pathway_analysis(object)
}
\arguments{
\item{object}{A \code{CASCAM} object.}
}
\description{
Pathway specific selection analysis
}
\examples{
\dontrun{
gene_info <- create_InformativeGenes(tumor_ct, tumor_label2, "ILC")
CASCAM_eg <- create_CASCAM(tumor_aligned, tumor_label, camod_aligned, gene_info)
CASCAM_eg <- sda_model(CASCAM_eg)
CASCAM_eg <- genome_selection(CASCAM_eg)
CASCAM_eg <- genome_selection_visualize(CASCAM_eg)
CASCAM_eg <- pathway_analysis(CASCAM_eg)
}
}
