% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_pathway_analysis.R
\name{pathway_specific_ridgeline}
\alias{pathway_specific_ridgeline}
\title{Pathway specific gene expression ridgeline}
\usage{
pathway_specific_ridgeline(object, interested_cells, pathway_name)
}
\arguments{
\item{object}{A \code{CASCAM} object.}

\item{interested_cells}{A vector of interested cell lines.}

\item{pathway_name}{The interested pathway, which should be within the KEGG and Hallmark pathways.}
}
\value{

}
\description{
Pathway specific gene expression ridgeline
}
