% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_shiny_app.R
\name{run_CASCAM_app}
\alias{run_CASCAM_app}
\title{Shiny app for CASCAM}
\usage{
run_CASCAM_app()
}
\description{
This is the function to call a Shiny GUI app for data exploration. After preparing a CASCAM object, users can call this function to perform the
genome-wide selection and pathway specific analysis.
}
