% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_shiny_app.R
\name{run_GULL_app}
\alias{run_GULL_app}
\title{Shiny app for GULL}
\usage{
run_GULL_app()
}
\value{

}
\description{
This is the function to call a Shiny GUI app for data exploration. After preparing a GULL object, users can call this function to perform the
genome-wide selection and pathway specific analysis.
}
