% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_genome_analysis.R
\name{sda_model}
\alias{sda_model}
\title{SDA model trained by tumor gene expression}
\usage{
sda_model(object, stop = NULL, lambda = NULL)
}
\arguments{
\item{object}{A \code{CASCAM} object.}

\item{stop}{A negative value with its absolute value representing the number of genes selected for model training.}

\item{lambda}{A parameter for the L2-norm for elastic net regression.}
}
\value{
A \code{CASCAM} object with sda_model, cell_norm_data, tumor_norm_data, tumor_sda_project, and cell_sda_project slots.
}
\description{
This function trains a sparse discriminant analysis model using tumor gene expression data. The parameters can be self-determined or
determined by the cross-validation. If cross-validation is prefered, users can refer to the function \link{sda_model_cv}.
}
\examples{

}
\references{
Clemmensen, Line, Trevor Hastie, Daniela Witten, and Bjarne Ersbøll. 2011. “Sparse Discriminant Analysis.”
 Technometrics: A Journal of Statistics for the Physical, Chemical, and Engineering Sciences 53 (4): 406–13. https://doi.org/10.1198/tech.2011.08118.
}
