% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_genome_analysis.R
\name{sda_model_cv}
\alias{sda_model_cv}
\title{SDA model trained by tumor gene expression}
\usage{
sda_model_cv(
  object,
  stop_vector = NULL,
  lambda_vector = NULL,
  parallel_cores = 1
)
}
\arguments{
\item{object}{A \code{CASCAM} object.}

\item{stop_vector}{A vector of negative values with its absolute value representing the number of genes selected for model training.}

\item{lambda_vector}{A vector of values for the L2-norm for elastic net regression.}

\item{parallel_cores}{Number of cores for the parallel computation.}
}
\value{
A \code{CASCAM} object with sda_model, camod_norm_data, tumor_norm_data, tumor_sda_project, and camod_sda_project slots.
}
\description{
This function trains a sparse discriminant analysis model using tumor gene expression data with cross-validation for parameter selection.
 Users can refer to the function \link{sda_model} if parameters are known.
}
\examples{
\dontrun{
gene_info <- create_InformativeGenes(tumor_ct, tumor_label2, "ILC")
CASCAM_eg <- create_CASCAM(tumor_aligned, tumor_label, camod_aligned, gene_info)
CASCAM_eg <- sda_model_cv(CASCAM_eg)
}

}
\references{
Clemmensen, Line, Trevor Hastie, Daniela Witten, and Bjarne Ersbøll. 2011. “Sparse Discriminant Analysis.”
 Technometrics: A Journal of Statistics for the Physical, Chemical, and Engineering Sciences 53 (4): 406–13. https://doi.org/10.1198/tech.2011.08118.
}
