% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_pathway_analysis.R
\name{see_pathview}
\alias{see_pathview}
\title{Pathview while not saving as files}
\usage{
see_pathview(..., save_image = FALSE)
}
\arguments{
\item{save_image}{}
}
\description{
Pathview while not saving as files
}
\keyword{internal}
