% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_genome_analysis.R
\name{genome_selection}
\alias{genome_selection}
\title{Genome-wide pre-selection analysis}
\usage{
genome_selection(object, R = 1000)
}
\arguments{
\item{object}{A \code{GULL} object.}

\item{R}{The number of bootstrap replicates. R = 1000 by default.}
}
\value{
A \code{GULL} object with genome-wide pre-selection related slots.
}
\description{
Genome-wide pre-selection analysis
}
