% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_pathway_analysis.R
\name{pathview_analysis}
\alias{pathview_analysis}
\title{Pathview of the interested cell in interested pathway}
\usage{
pathview_analysis(object, interested_cell, pathway_name)
}
\arguments{
\item{object}{A \code{GULL} object.}

\item{interested_cell}{A string of one interested cell line.}

\item{pathway_name}{The interested pathway, which should be within the KEGG and Hallmark pathways.}
}
\description{
Pathview of the interested cell in interested pathway
}
