% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_pathway_analysis.R
\name{pathway_congruence_heatmap}
\alias{pathway_congruence_heatmap}
\title{Pathway congruence heatmap}
\usage{
pathway_congruence_heatmap(object)
}
\arguments{
\item{object}{A \code{GULL} object.}
}
\value{

}
\description{
A heatmap showing the congruence between the cell lines and tumors for detailed pathways.
}
