% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_pathway_analysis.R
\name{pathway_specific_heatmap}
\alias{pathway_specific_heatmap}
\title{Pathway specific heatmap}
\usage{
pathway_specific_heatmap(object, pathway_name)
}
\arguments{
\item{object}{A \code{GULL} object.}

\item{pathway_name}{The interested pathway, which should be within the KEGG and Hallmark pathways.}
}
\value{

}
\description{
Pathway specific heatmap
}
